# packages/ms-gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Backend: O PokéCenter de Pallet Town",
    "version": "1.0.0",
    "title": "Swagger specification",
    "contact": {
      "email": "paulofellix@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Localhost (uses test data)"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "Status route, so we can check if server is alive",
        "tags": ["Status"]
      }
    },
    "/health": {
      "get": {
        "description": "Status route, so we can check if server is alive",
        "tags": ["Status"]
      }
    },
    "/cliente/{id}": {
      "get": {
        "tags": ["Cliente"],
        "description": "Retorna informações pertinentes o cliente (e o saldo atual)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Informações do cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "telefone": {
                      "type": "string"
                    },
                    "saldo": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cliente": {
      "post": {
        "tags": ["Cliente"],
        "description": "Cadastra informações pertinentes o cliente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "nome": {
                    "type": "string"
                  },
                  "telefone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente cadastrado com sucesso"
          }
        }
      }
    },
    "/saldo": {
      "get": {
        "tags": ["Financeiro"],
        "description": "Retorna o saldo de um cliente",
        "parameters": [
          {
            "name": "id_do_cliente",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saldo do cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "saldo": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transacao": {
      "post": {
        "tags": ["Financeiro"],
        "description": "Adiciona uma transação para um cliente (crédito ou débito)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_do_cliente": {
                    "type": "integer"
                  },
                  "valor": {
                    "type": "number"
                  },
                  "tipo": {
                    "type": "string",
                    "enum": ["credito", "debito"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transação adicionada com sucesso"
          }
        }
      }
    }
  }
}